@@ -1,15 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
-    buildToolsVersion "30.0.1"
+    compileSdkVersion 33
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
         applicationId "com.vibhunorby.totalpaisa"
-        minSdkVersion 21
-        targetSdkVersion 30
-        versionCode 17
-        versionName "1.0.16"
+        minSdkVersion 23
+        targetSdkVersion 33
+        versionCode 27
+        versionName "28.04.23.27"
+        // 27 - In app purchase added, Ui Enhancements, version format changed  and many more.
         multiDexEnabled true
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -28,22 +29,23 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'com.google.android.material:material:1.6.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-    implementation 'com.google.android.gms:play-services-ads:20.2.0'
+    implementation project(path: ':nativetemplates')
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'com.google.android.gms:play-services-ads:22.0.0'
     implementation 'com.android.support:multidex:1.0.3'
-    implementation 'androidx.biometric:biometric:1.0.1'
-    implementation 'com.google.android.play:core:1.8.0'
+    implementation 'androidx.biometric:biometric:1.1.0'
+    implementation 'com.google.android.play:core:1.10.3'
     implementation 'com.vorlonsoft:androidrate:1.0.8'
     implementation 'com.github.numerative:Five-Star-Me:2.0.0'
-    implementation 'com.anjlab.android.iab.v3:library:1.0.44'
-//    def billing_version = '4.0.0'
-//    implementation "com.android.billingclient:billing:$billing_version"
+
+    def billing_version = "5.2.0"
+    implementation "com.android.billingclient:billing:$billing_version"
 
 
 
