@@ -3,16 +3,20 @@ package com.vibhunorby.totalpaisa;
 import android.content.Context;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -26,11 +30,14 @@ import static com.vibhunorby.totalpaisa.MainActivity.save_button_pressed;
 
 public class HistoryFragment extends Fragment {
 
+    Prefs prefs;
     private ArrayList<CurrencyModal> currencyModalArrayList;
     private DBHandler dbHandler;
     private CurrencyRVAdapter currencyRVAdapter;
     private RecyclerView currencyRV;
     private LinearLayout textViewResultAmountLayout;
+    static ImageView imageViewNotFound;
+    static boolean isRecordVisible = false;
 
 
 
@@ -48,24 +55,26 @@ public class HistoryFragment extends Fragment {
         View view = inflater.inflate(R.layout.row_history, container, false);
         setHasOptionsMenu(true);
 
+        prefs = new Prefs(requireContext());
         currencyRV = view.findViewById(R.id.RVCurrencies);
-
+        imageViewNotFound = view.findViewById(R.id.image_view_not_found);
 
         return view;
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
-
-
-    }
+//    @Override
+//    public void onResume() {
+//        super.onResume();
+//
+//
+//    }
 
     public void setUserVisibleHint(boolean isVisibleToUser) {
 
         super.setUserVisibleHint(isVisibleToUser);
         if(isVisibleToUser) {
 
+            isRecordVisible = true;
 
 //            I am using this condition to not to refresh history fragment each and everytime user visit
 //                    refresh only first time the app starts and when any data saved.
@@ -80,7 +89,7 @@ public class HistoryFragment extends Fragment {
 
                 // on below line passing our array lost to our adapter class.
                 currencyRVAdapter = new CurrencyRVAdapter(currencyModalArrayList, getContext());
-
+                currencyRVAdapter.setHasStableIds(true);
 
 
 
@@ -88,9 +97,24 @@ public class HistoryFragment extends Fragment {
                 LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getContext(), RecyclerView.VERTICAL, false);
                 currencyRV.setLayoutManager(linearLayoutManager);
 
+                currencyRV.setHasFixedSize(true);
                 // setting our adapter to recycler view.
 
-                currencyRV.setAdapter(currencyRVAdapter);
+
+
+            // this delay is used to slow down the refresh lag (switching tab fragments lag)
+//            this lag is not found in Easy loan interest calculator, Reason is not found
+            final Handler handler = new Handler(Looper.getMainLooper());
+            handler.postDelayed(new Runnable() {
+                @Override
+                public void run() {
+
+
+                    currencyRV.setAdapter(currencyRVAdapter);
+
+                }
+            }, 300);
+
 
         }
 
@@ -101,9 +125,8 @@ public class HistoryFragment extends Fragment {
     public void onPrepareOptionsMenu(@NonNull Menu menu) {
         menu.findItem(R.id.RefreshButton).setVisible(false);
         menu.findItem(R.id.ShareButton).setVisible(false);
-        if(adRemoved){
-            menu.findItem(R.id.RemoveAd).setVisible(false);
-        } else {
+
+        if (prefs.isRemoveAd()) {
             menu.findItem(R.id.RemoveAd).setVisible(false);
         }
 
